fix: preserve bare ffmpeg PATH lookup in hardware-encoder probe
Commit 27a5bc3 wrapped the probe's executable in path.resolve() to address a
Semgrep detect-child-process finding. That broke PATH discovery: resolveFfmpegPath()
intentionally returns the bare command `ffmpeg` as its last fallback, and
path.resolve('ffmpeg') rewrites it to <cwd>/ffmpeg, so the "Faster" export mode's
hardware-encoder probe failed with ENOENT and silently fell back to libx264 on any
system relying on a PATH-installed FFmpeg. The real export spawn and availability
probe were untouched, so the regression was invisible in normal use.

- Extract runProbeProcess into main/ipc/ffmpegProbe.js so it is importable by the
  Node test runner (lyricVideoExport.js top-level-imports electron).
- Drop path.resolve(); keep shell:false explicit and args as a separate array. The
  executable is resolved internally at the resolveFfmpegPath() trust boundary, never
  from the export IPC request, so the finding is suppressed with a justified nosemgrep.
- Document resolveFfmpegPath() as the single resolution trust boundary.
- Add tests/ffmpeg-probe-path.test.js proving a bare ffmpeg on PATH is usable by the
  probe (and that a path.resolve()d command fails), guarding the regression.
